% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ae}
\alias{ae}
\title{Adverse event data}
\format{
\subsection{\code{who}}{

A tibble with 1191 rows and 35 columns:
\describe{
\item{STUDYID}{Study ID}
\item{DOMAIN}{Data domain}
...
}
}
}
\source{
\url{https://github.com/cdisc-org/sdtm-adam-pilot-project/raw/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/ae.xpt}
}
\usage{
ae
}
\description{
TBD
}
\keyword{datasets}
