% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merged_data}
\alias{merged_data}
\title{Merged data}
\format{
\subsection{\code{who}}{

A tibble with 59580 rows and 46 columns:
\describe{
\item{STUDYID}{Study ID}
\item{DOMAIN}{Data domain}
...
}
}
}
\source{
\url{https://github.com/cdisc-org/sdtm-adam-pilot-project/raw/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations}
}
\usage{
merged_data
}
\description{
raw_data$lab are merged to raw_data$demo by "STUDYID" and "USUBJID" via inner_join.
}
\keyword{datasets}
