% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_data}
\alias{raw_data}
\title{CDISC clinical data}
\format{
\subsection{\code{who}}{

A list of tibbles:
\describe{
\item{demo}{Demographic data}
\item{expo}{Exposure data}
\item{lab}{Laboratory data}
\item{ae}{Adverse event data}
}
}
}
\source{
\url{https://github.com/cdisc-org/sdtm-adam-pilot-project/raw/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations}
}
\usage{
raw_data
}
\description{
TBD
}
\keyword{datasets}
